+++ src/core/Pile.gd
@@ -3,6 +3,9 @@
 class_name Pile
 extends  CardContainer
 
+# If this is set to true, cards on this stack will be placed face-up.
+# Otherwise they will be placed face-down.
+export var faceup_cards := false
 
 # Called when the node enters the scene tree for the first time.
 func _ready():
@@ -20,11 +23,6 @@ func _process(_delta) -> void:
 	for obj in $ViewPopup/CardView.get_children():
 		if not obj.get_child_count():
 			obj.queue_free()
-		# This is needed for the instance where a player tries to drag a card from the gid
-		# to the same CardContainer
-		elif obj.get_child(0).modulate[3] != 1:
-				obj.get_child(0).modulate[3] = 1
-				obj.get_child(0).scale = Vector2(0.75,0.75)
 	# We make sure to adjust our popup if cards were removed from it while it's open
 	$ViewPopup.set_as_minsize()
 
@@ -32,6 +30,9 @@ func _process(_delta) -> void:
 # Populates the popup view window with all the cards in the deck
 # then displays it
 func _on_View_Button_pressed() -> void:
+	# We prevent the button from being pressed twice while the popup is open
+	# as it will bug-out
+	$Control/ManipulationButtons.visible = false
 	# We set the size of the grid to hold slightly scaled-down cards
 	for card in get_all_cards(false):
 		# We remove the card to rehost it in the popup grid container
@@ -70,9 +71,15 @@ func _on_ViewPopup_popup_hide() -> void:
 			# We need to remember that cards in piles should be left invisible
 			# and at default scale
 			card.scale = Vector2(1,1)
-			card.modulate[3] = 0
-			card.is_faceup = false
-
+			# We ensure that that closing popups reset card faceup to
+			# whatever is the default for the pile
+			if card.is_faceup != faceup_cards:
+				card.set_is_faceup(faceup_cards,true)
+			card.state = card.IN_PILE
+	reorganize_stack()
+	# We prevent the button from being pressed twice while the popup is open
+	# as it will bug-out
+	$Control/ManipulationButtons.visible = true
 
 # Overrides the built-in add_child() method,
 # To make sure the control node is set to be the last one among siblings.
@@ -90,6 +97,13 @@ func add_child(node, _legible_unique_name=false) -> void:
 		# we move them automatically to the viewpopup grid.
 		_slot_card_into_popup(node)
 
+func reorganize_stack() -> void:
+	for c in get_all_cards():
+		if c.position != Vector2(0.5 * get_card_index(c),
+				-1 * get_card_index(c)):
+			c.position = Vector2(0.5 * get_card_index(c),
+					-1 * get_card_index(c))
+					
 
 # Override the godot builtin move_child() method,
 # to make sure the $Control node is always drawn on top of Card nodes
@@ -120,7 +134,10 @@ func get_top_card() -> Card:
 	# prevent from trying to retrieve more cards
 	# than are in our deck and crashing godot.
 	if get_card_count():
-		card = get_card(0)
+		# Counter intuitively, the "top" card in the pile
+		# is the last node in the node hierarchy, so
+		# to retrieve the last card placed, we choose the last index
+		card = get_all_cards().back()
 	return card # Returning the card object for unit testing
 
 
@@ -130,14 +147,20 @@ func get_bottom_card() -> Card:
 	# prevent from trying to retrieve more cards
 	# than are in our deck and crashing godot.
 	if get_card_count():
-		card = get_card(get_card_count() - 1)
+		# Counter intuitively, the "bottom" card in the pile
+		# as it appears on screen, is the first node in the node hierarchy, so
+		# to retrieve the last c
+		card = get_card(0)
 	return card # Returning the card object for unit testing
 
+
+func get_stack_position(card: Card) -> Vector2:
+	return Vector2(0.5 * get_card_index(card), -1 * get_card_index(card))
 
 # Prepares a Card object to be added to the popup grid
 func _slot_card_into_popup(card: Card) -> void:
 	# We need to make the cards visible as they're by default invisible in piles
-	card.modulate[3] = 1
+	#card.modulate[3] = 1
 	# We also scale-down the cards to be able to see more at the same time.
 	# We need to do it before we add the card object to the control temp,
 	# otherwise it will default to the pre-scaled size
@@ -154,5 +177,12 @@ func _slot_card_into_popup(card: Card) -> void:
 	$ViewPopup/CardView.add_child(card_slot)
 	# Finally, the card is added to the temporary control node parent.
 	card_slot.add_child(card)
-	card.is_faceup = true
+	# warning-ignore:return_value_discarded
+	card.set_is_faceup(true,true)
+	card.position = Vector2(0,0)
+	card.state = card.IN_POPUP
 
+# Randomly rearranges the order of the Card nodes.
+func shuffle_cards() -> void:
+	.shuffle_cards()
+	reorganize_stack()