+++ tests/integration/TestVars.gd
@@ -1,5 +1,0 @@
-class_name TestVars
-extends Reference
-
-const mainScene = preload("res://src/core/Main.tscn")
-const boardScene = preload("res://src/custom/Board.tscn")