+++ changelog.md
@@ -5,6 +5,9 @@
 * Added new card back
 * Made the card back glow-pulse 
 * Can add or remove tokens from cards
+* Piles now display the size of the card stack
+* Cards in Piles are now places face-up or facedown depending on the pile config
+
 
 ## 0.10